Update Greater than, lower than Assignment Q2.py

+++ 4.9.25/Greater than, lower than Assignment Q2.py
@@ -2,7 +2,7 @@
 #Date: 4/9/25
 #Description: 2. Ask the user to enter a number under 20. If they enter a number that is 20 or over, display the message 'Too High'. Otherwise, display the message 'Thank you: '.
 
-number = int(input('Enter a number under 20: '))
+number = float(input('Enter a number under 20: '))
 
 if (number<20):
     print(f'{number} is under 20')
@@ -12,3 +12,4 @@ if (number>20):
 
 if (number == 20):
     print(f'{number} is too high')
+
